Servant.HAL

Description

Data types based on HAL (Hypertext Application Language) IETF draft: https://tools.ietf.org/html/draft-kelly-json-hal-08

Documentation

>>> :set -XDeriveGeneric
>>> :set -XDataKinds
>>> :set -XTypeOperators
>>> :set -XFlexibleInstances
>>> :set -XMultiParamTypeClasses
>>> import Servant.API
>>> import Data.Aeson
>>> import GHC.Generics

data HALLink

A HAL+JSON link object. Must contain a URI, may have other properties as specified in the spec.

Constructors

HALLink

Fields

_href :: URI

_templated :: Maybe Bool

_type :: Maybe MediaType

_deprecation :: Maybe URI

_name :: Maybe String

_profile :: Maybe URI

_title :: Maybe String

_hreflang :: Maybe String

HALLinks [HALLink]

data HAL a

A HAL+JSON resource. May contain links or embed other resources.

Constructors

HAL

Fields

resource :: a

_links :: HashMap Text HALLink

_embedded :: HashMap Text (HAL Value)

HALArray [HAL Value]

class ToJSON a => ToHAL a where

Generate a HAL a representation from a.

Methods

toHAL :: a -> HAL a

data HALJSON

HAL+JSON content type.

emptyLink :: HALLink

An empty HALLink.

uriFromString :: String -> URI

Turn a string into a URI or error out (used for URI literals).

toHALLink :: (IsElem endpoint api, HasLink endpoint) => Proxy api -> Proxy endpoint -> (MkLink endpoint -> Link) -> HALLink

Type-safe creation of a HALLink.

You must supply a function f that translates the output of safeLink into a Link. If the endpoint has no Capture or other components that take an argument, this function can be id (see toHALLink').

toHALLink' :: (IsElem endpoint api, HasLink endpoint, MkLink endpoint ~ Link) => Proxy api -> Proxy endpoint -> HALLink

Type-safe creation of a HALLink for an endpoint with no arguments.

The endpoint must not contain any Capture or other components needing input.

halWithSelf :: (ToJSON a, IsElem endpoint api, HasLink endpoint) => Proxy api -> Proxy endpoint -> (MkLink endpoint -> Link) -> a -> HAL a

Generate a HAL a with a "self" link.

As with toHALLink, you must supply a function from 'MkLink endpoint' to Link.

halWithSelfID :: (ToJSON a, IsElem endpoint api, HasLink endpoint, MkLink endpoint ~ (i -> Link)) => Proxy api -> Proxy endpoint -> (a -> i) -> a -> HAL a

Generate a HAL a with a "self" link produced by an ID getter.

Example:

>>> data User = User { userID :: Int, name :: String } deriving Generic
>>> instance ToJSON User
>>> type UserIndex = "users" :> Get '[HALJSON] [User]
>>> type UserShow = "users" :> Capture "id" Int :> Get '[HALJSON] User
>>> type UserAPI = UserIndex :<|> UserShow
>>> :{
instance ToHAL User where
  toHAL = halWithSelfID (Proxy :: Proxy UserAPI) (Proxy :: Proxy UserShow) userID
:}

halWithSelfRoute :: (ToJSON a, IsElem endpoint api, HasLink endpoint, MkLink endpoint ~ Link) => Proxy api -> Proxy endpoint -> a -> HAL a

Generate a HAL a with a "self" link produced by a simple route.

As with toHALLink', this presumes a route with no Captures or other components needing input.

Example:

>>> data Homepage = Homepage { homeData :: String } deriving Generic
>>> instance ToJSON Homepage
>>> type HomeRoute = Get '[HALJSON] Homepage
>>> type HomeAPI = HomeRoute
>>> :{
instance ToHAL Homepage where
  toHAL = halWithSelfRoute (Proxy :: Proxy HomeAPI) (Proxy :: Proxy HomeRoute)
:}

Orphan instances

ToJSON MediaType

Methods

toJSON :: MediaType -> Value

toEncoding :: MediaType -> Encoding

toJSONList :: [MediaType] -> Value

toEncodingList :: [MediaType] -> Encoding

ToJSON URI

Methods

toJSON :: URI -> Value

toEncoding :: URI -> Encoding

toJSONList :: [URI] -> Value

toEncodingList :: [URI] -> Encoding

FromJSON MediaType

Methods

parseJSON :: Value -> Parser MediaType

parseJSONList :: Value -> Parser [MediaType]

FromJSON URI

Methods

parseJSON :: Value -> Parser URI

parseJSONList :: Value -> Parser [URI]