Servant.StateGraph.Graph.Links

Description

Typeclass for specifying the links associated with a particular API.

Documentation

>>> :set -XDataKinds
>>> :set -XTypeOperators
>>> :set -XFlexibleInstances
>>> :set -XMultiParamTypeClasses
>>> import Servant.API

apiLinkToURI :: ApiLink -> URI

Convert an ApiLink to a URI situated at the root.

type (:=>) a b = (Proxy a, Proxy b)

Infix operator for expressing a link between two endpoints.

edgeFrom :: (Proxy a, Proxy b)

A convenient value to be used with linkFor, as in:

>>> type Counter = "counter" :> Get '[JSON] Int
>>> type CounterIncrement = "counter" :> "inc" :> Post '[JSON] Int
>>> let edge = edgeFrom :: Counter :=> CounterIncrement
>>> edge
(Proxy,Proxy)
>>> :{
instance LinksFor (Counter :<|> CounterIncrement) where
  linksFor api = [ linkFor api edge NormalNode ]
:}

class LinksFor api where

Define the links that lead to this return type.

>>> data User
>>> type UserIndex = "users" :> Get '[JSON] [User]
>>> type UserShow = "users" :> Capture "id" Int :> Get '[JSON] User
>>> type API = UserIndex :<|> UserShow
>>> :{
instance LinksFor API where
  linksFor api = [ linkFor api (edgeFrom :: UserIndex :=> UserShow) NormalNode ]
:}

Methods

linksFor :: Proxy api -> [RichLink]

linkFor :: (IsElem source api, IsElem target api, HasRichEndpoint source, HasRichEndpoint target) => Proxy api -> (Proxy source, Proxy target) -> NodeType -> RichLink

nodeTypesFromLinks :: [RichLink] -> Map TypeRep NodeType

Extract the return types from a list of RichLinks.