Servant.StateGraph.Graph.RichEndpoint

Description

A RichEndpoint makes information about an endpoint type accessible at the term level.

Documentation

class HasRichEndpoint endpoint where

Extract the ApiLink, return type, and method from an endpoint.

Minimal complete definition

getRichEndpoint

Methods

getRichEndpoint :: Proxy endpoint -> RichEndpoint

data RichEndpoint

Contains information about an endpoint, including a link to reach it, its HTTP StdMethod, and its return type.

Constructors

RichEndpoint

Fields

_apiLink :: ApiLink

_endpointMethod :: StdMethod

_returnType :: TypeRep