Servant.StateGraph.Graph.Server

Description

The included graph viewer can be served directly from a Servant API.

Documentation

data HTML

type GraphViewAPI = Get '[HTML] ApiGraph

graphTemplate :: Template

graphApi :: Proxy GraphViewAPI

app :: ApiGraph -> Application

serveGraph :: ApiGraph -> Int -> IO ()

Serve up the graph viewer on the specified port

Orphan instances

ToMustache ApiGraph

Methods

toMustache :: ApiGraph -> Value

listToMustache :: [ApiGraph] -> Value