Servant.StateGraph.Graph.Types

Description

Core types for representing nodes and edges in the state graph.

Documentation

data Root

A marker for the Root node.

Graph Types

data NodeType

Types of a state node.

Constructors

NormalNode
An ordinary node.

TargetNode
A node representing a target state.

ErrorNode
A node representing an error state.

data ApiNode

An ApiNode contains the name of the state node (the type of its representation) and its NodeType (whether it is a normal, target, or error state).

Constructors

ApiNode

Fields

apiNodeName :: String

apiNodeType :: NodeType

data ApiEdge

An ApiEdge contains the name of the link (by default, the URI traversed) and its color: see verbColors for list.

Constructors

ApiEdge

Fields

apiEdgeName :: String

apiEdgeColor :: String

type ApiGraph = Gr ApiNode ApiEdge

Alias for a PatriciaTrie Gr graph.

data ApiLink

Like Link, but exposing its constructor for manipulation.

Constructors

ApiLink

Fields

_segments :: [String]

_queryParams :: [String]

segments :: Lens' ApiLink [String]

queryParams :: Lens' ApiLink [String]

emptyLink :: ApiLink

An empty ApiLink.

data RichEndpoint

Contains information about an endpoint, including a link to reach it, its HTTP StdMethod, and its return type.

Constructors

RichEndpoint

Fields

_apiLink :: ApiLink

_endpointMethod :: StdMethod

_returnType :: TypeRep

returnType :: Lens' RichEndpoint TypeRep

endpointMethod :: Lens' RichEndpoint StdMethod

apiLink :: Lens' RichEndpoint ApiLink

data RichLink

Stores the source and target RichEndpoints, the target URI, NodeType, and name of the link relation.

Constructors

RichLink

Fields

_linkSource :: RichEndpoint

_linkTarget :: RichEndpoint

_linkTargetURI :: URI

_linkTargetNodeType :: NodeType

_linkRel :: String

linkTargetURI :: Lens' RichLink URI

linkTargetNodeType :: Lens' RichLink NodeType

linkTarget :: Lens' RichLink RichEndpoint

linkSource :: Lens' RichLink RichEndpoint

linkRel :: Lens' RichLink String

data GraphStateData

State data for building up the graph.

Constructors

GraphStateData

Fields

_typesToNodeIds :: Map TypeRep Node

_nodeTypes :: Map TypeRep NodeType

_edgesQueue :: [RichLink]

_currentNodeId :: Int

_currentGraph :: ApiGraph

typesToNodeIds :: Lens' GraphStateData (Map TypeRep Node)

nodeTypes :: Lens' GraphStateData (Map TypeRep NodeType)

edgesQueue :: Lens' GraphStateData [RichLink]

currentNodeId :: Lens' GraphStateData Int

currentGraph :: Lens' GraphStateData ApiGraph